+++ lib/features/ai_chat/application/providers/ai_providers.g.dart
@@ -6,7 +6,7 @@ part of 'ai_providers.dart';
 // RiverpodGenerator
 // **************************************************************************
 
-String _$availableModelsHash() => r'4b70bfd8aff92ce730fe1fbea4cb096bc9b218b9';
+String _$availableModelsHash() => r'6bb0c3017dec2e049dbae96d033f16e024a51355';
 
 /// 所有可用的模型配置列表 Provider
 ///